Reorganized Component Folders
