+++ src/Components/Expenses/ExpenseItem.js
@@ -1,6 +1,6 @@
 import React from "react";
 import './ExpenseItem.css';
-import Card from './Card';
+import Card from '../UI/Card';
 import ExpenseDate from './ExpenseDate';
 
 function ExpenseItem(props) {
