+++ src/Components/Expenses/Expenses.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ExpenseItem from './ExpenseItem';
-import Card from './Card'
+import Card from '../UI/Card'
 import './Expenses.css'
 
 function Expenses(props) {
